switch from codes to commodity_list, use run_type arg
Explanation of why: Commodity was changed to a list to allow multiple commodities. this list has to be the same length as lambda_c_list.

As the code is now, it reads all the options in the trade folder as the commodities and runs the first. we previously handled this by only keeping data for commodities we were using, but the running from the list gives more flexibility.

The run_suffix is to write calibration and forecast runs to separate folders, but is optional.